// packages/cli/src/ui/types.d.ts
export declare enum StreamingState {
  Idle = 'idle',
  Responding = 'responding',
  WaitingForConfirmation = 'waiting_for_confirmation',
}
export declare enum GeminiEventType {
  Content = 'content',
  ToolCallRequest = 'tool_call_request',
}
export declare enum ToolCallStatus {
  Pending = 'Pending',
  Canceled = 'Canceled',
  Confirming = 'Confirming',
  Executing = 'Executing',
  Success = 'Success',
  Error = 'Error',
}